Extract app config core records
